requirements.txt
feedparser>=6.0.11
openai>=1.75.0
PyYAML>=6.0.2
requests>=2.32.3
